memory fix
